requirements.txt
langchain==0.0.154
PyPDF2==3.0.1
streamlit==1.18.1
qdrant-client==1.1.6
sentence-transformers==2.2.2
llama-cpp-python==0.1.48
pymongo==4.4.1
tika==1.18.0
python-dotenv==1.0.0
faiss-cpu==1.7.4
altair==4.2.2